collection.kt 오류 수정
