+++ app/src/main/res/values/dimens.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <dimen name="_200dp">200dp</dimen>
+    <dimen name="fab_margin">16dp</dimen>
 </resources>
